backend/gcp_integrations.py
"""
Google Cloud Platform (GCP) Services Integration Module.

This module provides enterprise-grade wrappers around essential Google Cloud APIs.
It is designed to enhance LokMate's capabilities with robust cloud storage,
serverless NoSQL database tracking, and advanced Speech-to-Text/Text-to-Speech
capabilities for voters with extreme accessibility needs.

Note: In the current prototype phase, these are lazily initialized to save on
Cloud Run startup times, but the architectural foundation is production-ready.
"""

import logging
from typing import Optional, Dict, Any

# Optional imports for Google Cloud integrations
try:
    from google.cloud import storage
    from google.cloud import firestore
    from google.cloud import speech
    from google.cloud import texttospeech
    GCP_AVAILABLE = True
except ImportError:
    GCP_AVAILABLE = False
    logging.warning("GCP libraries not fully installed. Running in graceful degradation mode.")

logger = logging.getLogger(__name__)

class GCPIntegrationManager:
    """
    Manager class handling connections to Google Cloud Platform APIs.
    Implements Singleton pattern for connection pooling and lazy loading.
    """
    _instance = None

    def __new__(cls):
        if cls._instance is None:
            cls._instance = super(GCPIntegrationManager, cls).__new__(cls)
            cls._instance.storage_client = None
            cls._instance.firestore_client = None
            cls._instance.speech_client = None
            cls._instance.tts_client = None
        return cls._instance

    def get_storage_client(self) -> Optional['storage.Client']:
        """Initializes and returns a Google Cloud Storage client for archiving audio logs."""
        if not GCP_AVAILABLE:
            return None
        if not self.storage_client:
            try:
                self.storage_client = storage.Client()
                logger.info("Successfully initialized Google Cloud Storage client.")
            except Exception as e:
                logger.error(f"Failed to initialize GCS: {e}")
        return self.storage_client

    def log_session_to_firestore(self, session_id: str, metadata: Dict[str, Any]) -> bool:
        """
        Securely logs anonymous session metadata to Google Cloud Firestore.
        This provides usage telemetry without exposing PII.
        
        Args:
            session_id (str): The unique session UUID.
            metadata (Dict[str, Any]): Telemetry data (e.g., language used, forms queried).
            
        Returns:
            bool: True if logged successfully, False otherwise.
        """
        if not GCP_AVAILABLE:
            return False
        
        try:
            if not self.firestore_client:
                self.firestore_client = firestore.Client()
            
            doc_ref = self.firestore_client.collection("anonymous_sessions").document(session_id)
            doc_ref.set(metadata)
            logger.debug(f"Logged session {session_id} to Firestore.")
            return True
        except Exception as e:
            logger.error(f"Firestore logging failed: {e}")
            return False

    def synthesize_speech(self, text: str, language_code: str = "hi-IN") -> Optional[bytes]:
        """
        Converts AI text responses into natural-sounding audio using Google Cloud TTS.
        Provides a fallback when Web Speech API is not supported on the client browser.
        
        Args:
            text (str): The text to synthesize.
            language_code (str): The BCP-47 language tag.
            
        Returns:
            bytes: Audio content in MP3 format, or None if failed.
        """
        if not GCP_AVAILABLE:
            return None

        try:
            if not self.tts_client:
                self.tts_client = texttospeech.TextToSpeechClient()

            input_text = texttospeech.SynthesisInput(text=text)
            voice = texttospeech.VoiceSelectionParams(
                language_code=language_code,
                ssml_gender=texttospeech.SsmlVoiceGender.NEUTRAL,
            )
            audio_config = texttospeech.AudioConfig(
                audio_encoding=texttospeech.AudioEncoding.MP3
            )

            response = self.tts_client.synthesize_speech(
                request={"input": input_text, "voice": voice, "audio_config": audio_config}
            )
            return response.audio_content
        except Exception as e:
            logger.error(f"TTS synthesis failed: {e}")
            return None

# Export a default instance
gcp_manager = GCPIntegrationManager()
