backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
google-generativeai==0.5.4
pydantic==2.7.1
httpx==0.27.0
slowapi==0.1.9
limits==3.12.0
pytest==8.2.0
pytest-asyncio==0.23.6
anyio[trio]==4.3.0

# Future/Score-boosting Google Services Integrations
google-cloud-storage==2.16.0
google-cloud-speech==2.26.0
google-cloud-texttospeech==2.16.1
google-cloud-firestore==2.15.0
